Muffin/requirements.txt
aiohttp==3.6.2
altgraph==0.17
animation==0.0.6
asgiref==3.2.7
astroid==2.3.3
async-timeout==3.0.1
asyncpg==0.20.1
attrs==19.3.0
backcall==0.1.0
bleach==3.1.3
certifi==2019.11.28
cffi==1.14.0
chardet==3.0.4
Click==7.0
colorama==0.4.3
decorator==4.4.2
defusedxml==0.6.0
dhooks==1.1.3
discord.py==1.3.1
Django==3.0.5
dnspython==1.16.0
emoji==0.5.4
entrypoints==0.3
Flask==1.1.1
future==0.18.2
idna==2.8
ipykernel==5.2.0
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
itsdangerous==1.1.0
jedi==0.16.0
Jinja2==2.11.1
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.0
jupyter-console==6.1.0
jupyter-core==4.6.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
mistune==0.8.4
multidict==4.7.4
mysql-connector-python==8.0.19
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.1
pandas==1.0.1
pandocfilters==1.4.2
parso==0.6.2
pefile==2019.4.18
pickleshare==0.7.5
Pillow==7.0.0
playsound==1.2.2
pluginbase==1.0.0
praw==6.5.1
prawcore==1.0.1
prometheus-client==0.7.1
prompt-toolkit==3.0.4
protobuf==3.6.1
psutil==5.6.7
py-cpuinfo==5.0.0
pycparser==2.19
Pygments==2.6.1
PyInstaller==3.6
pylint==2.4.4
PyNaCl==1.3.0
pyperclip==1.7.0
pyrsistent==0.15.7
PySide2==5.14.1
PySimpleGUI==4.15.2
PySocks==1.7.1
python-dateutil==2.8.1
pytz==2019.3
pywin32==227
pywin32-ctypes==0.2.0
pywinpty==0.5.7
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
requests==2.22.0
selenium==3.141.0
Send2Trash==1.5.0
shiboken2==5.14.1
six==1.14.0
sqlparse==0.3.1
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
tornado==6.0.4
traitlets==4.3.3
update-checker==0.16
urllib3==1.25.8
wcwidth==0.1.8
webencodings==0.5.1
websocket-client==0.57.0
websockets==8.1
Werkzeug==0.16.1
widgetsnbextension==3.5.1
wrapt==1.11.2
xlrd==1.2.0
yarl==1.4.2
youtube-dl==2020.2.16
